Move click category POI handler in scene file
